+++ test/e2e/tests/complex/nullable_types.2.0/expected_json/models/pet.dart
@@ -28,8 +28,9 @@ class Pet {
   final int? id;
   @JsonKey(includeIfNull: false)
   final Category? category;
-  final String name;
-  final List<String> photoUrls;
+  @JsonKey(includeIfNull: true)
+  final String? name;
+  final List<String?> photoUrls;
   @JsonKey(includeIfNull: false)
   final List<Tag>? tags;